+++ internal/elasticsearch/logstash/pipeline.go
@@ -67,17 +67,6 @@ func ResourceLogstashPipeline() *schema.Resource {
 			Required:    true,
 		},
 		"pipeline_metadata": {
-			Description:      "Optional metadata about the pipeline. This property will be removed in a future provider version. Use `metadata` field instead. (Deprecated)",
-			Type:             schema.TypeMap,
-			Optional:         true,
-			Deprecated:       "Manual changes to pipelines managed via Terraform will throw an error. Use `metadata` field instead.",
-			DiffSuppressFunc: utils.DiffJsonSuppress,
-			Elem: &schema.Schema{
-				Type:    schema.TypeString,
-				Default: nil,
-			},
-		},
-		"metadata": {
 			Description:      "Optional JSON metadata about the pipeline.",
 			Type:             schema.TypeString,
 			ValidateFunc:     validation.StringIsJSON,
@@ -212,15 +201,11 @@ func resourceLogstashPipelinePut(ctx context.Context, d *schema.ResourceData, me
 	logstashPipeline.LastModified = utils.FormatStrictDateTime(time.Now().UTC())
 	logstashPipeline.Pipeline = d.Get("pipeline").(string)
 
-	if len(d.Get("pipeline_metadata").(map[string]interface{})) != 0 {
-		logstashPipeline.Metadata = d.Get("pipeline_metadata").(map[string]interface{})
-	} else {
-		var metadata map[string]interface{}
-		if err := json.Unmarshal([]byte(d.Get("metadata").(string)), &metadata); err != nil {
-			return diag.FromErr(err)
-		}
-		logstashPipeline.Metadata = metadata
+	var pipelineMetadata map[string]interface{}
+	if err := json.Unmarshal([]byte(d.Get("pipeline_metadata").(string)), &pipelineMetadata); err != nil {
+		return diag.FromErr(err)
 	}
+	logstashPipeline.PipelineMetadata = pipelineMetadata
 
 	logstashPipeline.PipelineSettings = map[string]interface{}{}
 	if settings := utils.ExpandIndividuallyDefinedSettings(ctx, d, allSettingsKeys); len(settings) > 0 {
@@ -270,18 +255,12 @@ func resourceLogstashPipelineRead(ctx context.Context, d *schema.ResourceData, m
 		return diag.FromErr(err)
 	}
 
-	if len(d.Get("pipeline_metadata").(map[string]interface{})) != 0 {
-		if err := d.Set("pipeline_metadata", logstashPipeline.Metadata); err != nil {
-			return diag.FromErr(err)
-		}
-	} else {
-		metadata, err := json.Marshal(logstashPipeline.Metadata)
-		if err != nil {
-			return diag.FromErr(err)
-		}
-		if err := d.Set("metadata", string(metadata)); err != nil {
-			return diag.FromErr(err)
-		}
+	pipelineMetadata, err := json.Marshal(logstashPipeline.PipelineMetadata)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	if err := d.Set("pipeline_metadata", string(pipelineMetadata)); err != nil {
+		return diag.FromErr(err)
 	}
 
 	for key, typ := range allSettingsKeys {